Moved BackgroundTaskManager into separate module.
