+++ python/task_manager/__init__.py
@@ -8,4 +8,4 @@
 # agreement to the Shotgun Pipeline Toolkit Source Code License. All rights 
 # not expressly granted therein are reserved by Shotgun Software Inc.
 
-from .shotgun_data_retriever import ShotgunDataRetriever
+from .background_task_manager import BackgroundTaskManager
